Some code refactoring
