+++ ckanext/georchestra/commands/georchestra_ldap.py
@@ -6,9 +6,9 @@ from ckan.lib.cli import CkanCommand
 from ckan.plugins import toolkit
 from ckan.plugins.toolkit import config
 
-from ckanext.georchestra.commands.utils import ldap_utils
-from ckanext.georchestra.commands.utils import organizations as org_utils
-from ckanext.georchestra.commands.utils import users as user_utils
+from ckanext.georchestra.utils import ldap_utils
+from ckanext.georchestra.utils import organizations as org_utils
+from ckanext.georchestra.utils import users as user_utils
 
 
 log = logging.getLogger()
