+++ ckanext/georchestra/utils/ldap_utils.py
